Extract links scanner
